# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CatalogItem": {
        "properties": {
          "id": {
            "description": "Stable deterministic item identifier.",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "description": "Deterministic display name for the item.",
            "title": "Name",
            "type": "string"
          },
          "price": {
            "description": "Deterministic item price.",
            "title": "Price",
            "type": "number"
          }
        },
        "required": [
          "id",
          "name",
          "price"
        ],
        "title": "CatalogItem",
        "type": "object"
      },
      "CatalogPage": {
        "properties": {
          "attempt": {
            "description": "One-based request attempt for this run, scenario, and page.",
            "title": "Attempt",
            "type": "integer"
          },
          "items": {
            "description": "Items for the requested page.",
            "items": {
              "$ref": "#/components/schemas/CatalogItem"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "description": "One-based page that produced this response.",
            "title": "Page",
            "type": "integer"
          },
          "scenario": {
            "description": "Resolved scenario name.",
            "enum": [
              "success",
              "transient",
              "permanent",
              "slow",
              "resume",
              "dom-change",
              "duplicates",
              "selector-failure",
              "rate-limit"
            ],
            "title": "Scenario",
            "type": "string"
          },
          "total_pages": {
            "description": "Total pages exposed by the scenario.",
            "title": "Total Pages",
            "type": "integer"
          }
        },
        "required": [
          "page",
          "total_pages",
          "items",
          "scenario",
          "attempt"
        ],
        "title": "CatalogPage",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Resilient Browser Automation Test Stand",
    "version": "1.1.5"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/catalog": {
      "get": {
        "description": "Returns deterministic catalog data for retry, pagination, duplicate, rate-limit, delay, and checkpoint-recovery tests.",
        "operationId": "get_catalog_page",
        "parameters": [
          {
            "description": "Scenario to execute; inherits the active preset when omitted.",
            "in": "query",
            "name": "scenario",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "success",
                    "transient",
                    "permanent",
                    "slow",
                    "resume",
                    "dom-change",
                    "duplicates",
                    "selector-failure",
                    "rate-limit"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Scenario to execute; inherits the active preset when omitted.",
              "title": "Scenario"
            }
          },
          {
            "description": "Opaque key that isolates deterministic attempt counters between test runs.",
            "in": "query",
            "name": "run_id",
            "required": false,
            "schema": {
              "default": "manual",
              "description": "Opaque key that isolates deterministic attempt counters between test runs.",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "description": "Initial transient failures per page.",
            "in": "query",
            "name": "fail_for",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 10,
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Initial transient failures per page.",
              "title": "Fail For"
            }
          },
          {
            "description": "Response delay for the slow scenario, in milliseconds.",
            "in": "query",
            "name": "delay_ms",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 30000,
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Response delay for the slow scenario, in milliseconds.",
              "title": "Delay Ms"
            }
          },
          {
            "description": "Delay before each transient 503 response, in milliseconds.",
            "in": "query",
            "name": "failure_delay_ms",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 30000,
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Delay before each transient 503 response, in milliseconds.",
              "title": "Failure Delay Ms"
            }
          },
          {
            "description": "Permanently failing page in the resume scenario.",
            "in": "query",
            "name": "fail_page",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 20,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Permanently failing page in the resume scenario.",
              "title": "Fail Page"
            }
          },
          {
            "description": "Number of pages exposed by the catalog.",
            "in": "query",
            "name": "total_pages",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 20,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Number of pages exposed by the catalog.",
              "title": "Total Pages"
            }
          },
          {
            "description": "Initial HTTP 429 responses per page in the rate-limit scenario.",
            "in": "query",
            "name": "rate_limit_for",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 10,
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Initial HTTP 429 responses per page in the rate-limit scenario.",
              "title": "Rate Limit For"
            }
          },
          {
            "description": "Retry-After value for rate-limit responses, in seconds.",
            "in": "query",
            "name": "retry_after_seconds",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 300,
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Retry-After value for rate-limit responses, in seconds.",
              "title": "Retry After Seconds"
            }
          },
          {
            "description": "Browser locator target for the selector-failure scenario.",
            "in": "query",
            "name": "selector_failure_target",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "login_button",
                    "item",
                    "next_page"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Browser locator target for the selector-failure scenario.",
              "title": "Selector Failure Target"
            }
          },
          {
            "description": "Deterministic failure mode for the configured browser locator.",
            "in": "query",
            "name": "selector_failure_mode",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "missing",
                    "changed",
                    "multiple",
                    "delayed",
                    "hidden",
                    "disabled"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Deterministic failure mode for the configured browser locator.",
              "title": "Selector Failure Mode"
            }
          },
          {
            "description": "Catalog page where the locator failure is activated.",
            "in": "query",
            "name": "selector_failure_page",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 20,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Catalog page where the locator failure is activated.",
              "title": "Selector Failure Page"
            }
          },
          {
            "description": "Bounded delay before a configured locator appears.",
            "in": "query",
            "name": "selector_failure_delay_ms",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 5000,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bounded delay before a configured locator appears.",
              "title": "Selector Failure Delay Ms"
            }
          },
          {
            "description": "One-based catalog page to fetch.",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "description": "One-based catalog page to fetch.",
              "maximum": 20,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "description": "Rate-limit scenario response; includes the Retry-After header."
          },
          "500": {
            "description": "Permanent or checkpoint-resume scenario failure."
          },
          "503": {
            "description": "Transient scenario failure; includes the Retry-After header."
          }
        },
        "summary": "Fetch one deterministic catalog page"
      }
    },
...
  }
}